edge-node-components/db/requirements.txt
sqlalchemy
mariadb
alembic
paho-mqtt